requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
pytesseract==0.3.10
pandas==2.1.3
numpy==1.24.3
streamlit==1.28.1
plotly==5.17.0
python-dateutil==2.8.2
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6 